Fixed usability and selecting text files
I have fixed a usability problem, so instead of the user havibng to write how many items they would like to add, instead I have made it so if they want to add an item they'll just have to keep pressing the menu option that adds items.

I have now fixed it so that the user is able to select from a menu of text-files, and the one they select will be the one that loads.